Refactor AdminPanel to use reusable components
