+++ src/utils/productUtils.ts
@@ -1,0 +1,35 @@
+// Utility functions for product-related operations
+
+/**
+ * Determines if a product is new based on its creation date
+ * A product is considered "new" if it was created within the last 7 days
+ */
+export const isProductNew = (createdAt: string): boolean => {
+  const createdDate = new Date(createdAt);
+  const now = new Date();
+  const diffTime = Math.abs(now.getTime() - createdDate.getTime());
+  const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+  return diffDays <= 7;
+};
+
+/**
+ * Formats price for display in Brazilian Real format
+ */
+export const formatPrice = (price: number): string => {
+  return `R$ ${price.toFixed(2).replace('.', ',')}`;
+};
+
+/**
+ * Generates a fallback image URL for broken product images
+ */
+export const getFallbackImageUrl = (): string => {
+  return 'data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMjAwIiBoZWlnaHQ9IjIwMCIgeG1sbnM9Imh0dHA6Ly93d3cudzMub3JnLzIwMDAvc3ZnIj48cmVjdCB3aWR0aD0iMTAwJSIgaGVpZ2h0PSIxMDAlIiBmaWxsPSIjZjNmNGY2Ii8+PHRleHQgeD0iNTAlIiB5PSI1MCUiIGZvbnQtZmFtaWx5PSJBcmlhbCIgZm9udC1zaXplPSIxNCIgZmlsbD0iIzljYTNhZiIgdGV4dC1hbmNob3I9Im1pZGRsZSIgZHk9Ii4zZW0iPkltYWdlbSBuw6NvIGVuY29udHJhZGE8L3RleHQ+PC9zdmc+';
+};
+
+/**
+ * Handles image loading errors by setting a fallback image
+ */
+export const handleImageError = (event: React.SyntheticEvent<HTMLImageElement, Event>): void => {
+  const target = event.target as HTMLImageElement;
+  target.src = getFallbackImageUrl();
+};
